Use double precision and add RK4 integrator
Convert core physics types to double precision (BlackHole, Constants, RayConfig) and normalize units (c=G=1). Implement NewtonianAcceleration plus Euler and RK4 marching functions and use RK4 updates with a dynamic step size to improve trajectory integration accuracy. Increase window width to 1024, raise MAX_STEPS to 500, and downsample pixel iteration (step by 4) for performance. Update black hole initialization and replace simple ray-march with velocity-based propagation; add simple checkerboard background and black hole hit testing. Misc: adjust colors and reorder rendering/update loop as part of the changes.

+++ BlackHoleTracer/Headers/blackhole.h
@@ -2,16 +2,16 @@
 
 class BlackHole {
 public:
-    BlackHole(float mass, glm::vec3 position) : m_mass(mass), m_position(position) {
+    BlackHole(double mass, glm::dvec3 position) : m_mass(mass), m_position(position) {
         m_radius_s = 2 * Constants::G * m_mass / (Constants::c * Constants::c);
     }
 
-    float getMass() const { return m_mass; }
-    glm::vec3 getPosition() const { return m_position; }
-    float getRadius() const { return m_radius_s;}
+    double getMass() const { return m_mass; }
+    glm::dvec3 getPosition() const { return m_position; }
+    double getRadius() const { return m_radius_s;}
 
 private:
-    float m_mass;
-    float m_radius_s;
-    glm::vec3 m_position;
+    double m_mass;
+    double m_radius_s;
+    glm::dvec3 m_position;
 };
